create STAC item from TDS NcML response - untested due to missing STAC host
